requirements.txt
autogluon
pandas
openml
ruff
pyyaml
xlrd
openpyxl
seaborn
tueplots
latex
tqdm